Move tests to yaml_utils_test.go, changelog under 1.11.1
